Move from Domain to Usecases CommandExecutor and History and UndoManager
