--- rover/domain/src/main/kotlin/com/marsrover/NasaMarsRover/rover/domain/RoverCommandHistory.kt
@@ -1,16 +1,0 @@
-package com.marsrover.NasaMarsRover.rover.domain
-
-// глобальное хранилище истории команд для всех роверов
-class RoverCommandHistory {
-    private val commands: HashMap<RoverId, MutableList<RoverCommandExecutor.Command>> = hashMapOf()
-
-    fun addCommand(roverId: RoverId, command: RoverCommandExecutor.Command) {
-        commands.getOrPut(roverId) { mutableListOf() }.add(command)
-    }
-
-    fun getCommandHistory(roverId: RoverId): List<RoverCommandExecutor.Command>? {
-        return commands[roverId]
-    }
-}
-
-
